@@ -4,12 +4,11 @@ vg_config.py: Default configuration values all here (and only here), as well as 
 for reading and generating config files.
 
 """
-from __future__ import print_function
+
 import argparse, sys, os, os.path, errno, random, subprocess, shutil, itertools, glob, tarfile
 import doctest, re, json, collections, time, timeit
-import logging, logging.handlers, SocketServer, struct, socket, threading
+import logging, logging.handlers, struct, socket, threading
 import string
-import urlparse
 import getpass
 import pdb
 import textwrap
@@ -573,7 +572,7 @@ def generate_config(whole_genome = False):
     return whole_genome_config if whole_genome is True else default_config
 
 def make_opts_list(x_opts):
-    opts_list = filter(lambda a : len(a), x_opts.split(' '))
+    opts_list = list([a for a in x_opts.split(' ') if len(a)])
     # get rid of any -t or --threads while we're at it    
     for t in ['-t', '--threads']:
         if t in opts_list:
@@ -591,18 +590,18 @@ def apply_config_file_args(args):
     for x_opts in ['map_opts', 'call_opts', 'recall_opts', 'filter_opts', 'recall_filter_opts', 'genotype_opts',
                    'vcfeval_opts', 'sim_opts', 'bwa_opts', 'minimap2_opts', 'gcsa_opts', 'minimizer_opts', 'mpmap_opts',
                     'gaffe_opts', 'augment_opts', 'pack_opts', 'prune_opts']:
-        if x_opts in args.__dict__.keys() and type(args.__dict__[x_opts]) is str:
+        if x_opts in list(args.__dict__.keys()) and type(args.__dict__[x_opts]) is str:
             args.__dict__[x_opts] = make_opts_list(args.__dict__[x_opts])
 
     # do the same thing for more_mpmap_opts which is a list of strings
-    if 'more_mpmap_opts' in args.__dict__.keys() and args.__dict__['more_mpmap_opts']:
+    if 'more_mpmap_opts' in list(args.__dict__.keys()) and args.__dict__['more_mpmap_opts']:
         for i, m_opts in enumerate(args.__dict__['more_mpmap_opts']):
-            if isinstance(m_opts, basestring):
+            if isinstance(m_opts, str):
                 args.__dict__['more_mpmap_opts'][i] = make_opts_list(m_opts)
 
     # If no config file given, we generate a default one
-    wg_config = args.__dict__.has_key('whole_genome_config') and args.whole_genome_config
-    if not args.__dict__.has_key('config') or args.config is None:
+    wg_config = 'whole_genome_config' in list(args.__dict__.keys()) and args.whole_genome_config
+    if 'config' not in list(args.__dict__.keys()) or args.config is None:
         config = generate_config(whole_genome = wg_config)
     else:
         if wg_config:
@@ -613,7 +612,7 @@ def apply_config_file_args(args):
             config = conf.read()
                 
     # Parse config
-    parsed_config = {x.replace('-', '_'): y for x, y in yaml.safe_load(config).iteritems()}
+    parsed_config = {x.replace('-', '_'): y for x, y in list(yaml.safe_load(config).items())}
     if 'prune_opts_2' in parsed_config:
         raise RuntimeError('prune-opts-2 from config no longer supported')
     options = argparse.Namespace(**parsed_config)
@@ -624,7 +623,7 @@ def apply_config_file_args(args):
     for args_key in args.__dict__:
         # Add in missing program arguments to config option list and
         # overwrite config options with corresponding options that are not None in program arguments
-        if (args.__dict__[args_key] is not None) or (args_key not in  options.__dict__.keys()):
+        if (args.__dict__[args_key] is not None) or (args_key not in  list(options.__dict__.keys())):
             options.__dict__[args_key] = args.__dict__[args_key]
             
     return options
