@@ -1,15 +1,14 @@
-#!/usr/bin/env python2.7
+#!/usr/bin/env python
 """
 vg_msga.py: use vg's banded aligner to align contigs into the graph.  By associating BED regions with each
 contig, we can parallelize by chromosome. 
 
 """
-from __future__ import print_function
+
 import argparse, sys, os, os.path, errno, random, subprocess, shutil, itertools, glob, tarfile
 import doctest, re, json, collections, time, timeit
-import logging, logging.handlers, SocketServer, struct, socket, threading
+import logging, logging.handlers, struct, socket, threading
 import string
-import urlparse
 import getpass
 import pdb
 import gzip
@@ -128,7 +127,7 @@ def run_msga(job, context, graph_name, graph_id, fasta_id, target_regions_id, ch
         if target_regions_id and chrom:
             fasta_subset_path = os.path.join(work_dir, '{}.fa'.format(chrom))
             subset_count = 0
-            with open(fasta_subset_path, 'w') as subset_file:
+            with open(fasta_subset_path, 'wb') as subset_file:
                 for target_region in target_regions:
                     toks = line.strip().split('\t')
                     context.runner.call(job, ['samtools', 'faidx', os.path.basename(fasta_path),
